+++ src/components/zshrc/aliases/git.sh
@@ -8,4 +8,4 @@ alias gcwhatever='gaa && gc -m "Whatever $(date)" && ggpush'
 alias ggauthwill='git commit --amend --author="Will Gibson <will@biffbangpow.com>"'
 alias glog="git log --branches --tags --graph --oneline --decorate"
 alias ghist="git log --pretty=format:\"%h %ad | %s%d %an\" --graph --date=short"
-alias git-delete-all-local-branches-except-master-develop-and-current='git branch | grep -v "master" | grep -v "develop" | xargs git branch -D'
+alias git-delete-all-local-branches-except-main-develop-and-current='git branch | grep -v "master" | grep -v "main" | grep -v "develop" | xargs git branch -D'