+++ src/setup.sh
@@ -1,31 +1,45 @@
 #!/bin/bash
 
+set -e
+
 # shellcheck disable=SC2164
 basePath="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
 
+# In case there's something essential in there we hand rolled
 cp ~/.zshrc "${HOME}/.zshrc.backup.$(date)"
 
 # shellcheck disable=SC1090
-source "${basePath}/components/install_commands.sh"
+source "${basePath}/components/commands/docker.sh"
+source "${basePath}/components/commands/git.sh"
+source "${basePath}/components/commands/miscellaneous.sh"
+source "${basePath}/components/commands/nodejs.sh"
+source "${basePath}/components/commands/php.sh"
+source "${basePath}/components/commands/pull_latest.sh"
+source "${basePath}/components/commands/zshrc.sh"
 
 # Preflight checks
-if docker version | grep --quiet "Server: Docker Engine"; then
-    echo "Error: Please quit Docker before running the setup."
-    exit 1
-fi
+ensure_docker_not_running
+ensure_git_name_and_email_are_set
 
 append_to_zshrc "# The rest should have been added by laptop-setup..."
 
-ensure_git_name_and_email_are_set
+pull_latest_laptop_setup_code
 
-ensure_laptop_setup_is_up_to_date
+set +e && ensure_homebrew_is_installed_and_up_to_date && set -e
 
-ensure_homebrew_is_installed_and_up_to_date
-
+# Terminal
 brew reinstall iterm2
+ensure_zsh_is_installed
+brew reinstall zsh zsh-completions
+ensure_ohmyzsh_is_installed
+ensure_zsh_autosuggestions_are_installed
+ensure_zsh_correction_is_used
+ensure_zsh_completion_waiting_dots_are_used
+ensure_correct_ohmyzsh_theme_is_used "${basePath}/components/ohmyzsh/willgibson.zsh-theme" "willgibson"
+update_file_line_in_situ ~/.zshrc 'plugins=(git)' 'plugins=(docker git zsh-autosuggestions)'
 
-ensure_zsh_and_zsh_completions_are_installed
-
+# Git
+ensure_git_name_and_email_env_vars_are_exported_in_zshrc
 brew reinstall git
 git config --global pull.ff only
 
@@ -37,12 +51,11 @@ brew reinstall pinentry-mac
 ensure_php_is_installed
 brew reinstall composer
 
-# Node.js
-brew reinstall nvm
-mkdir ~/.nvm
-append_to_zshrc "source ${basePath}/components/zshrc/nvm_switcher.sh"
+# NVM & Node.js
+ensure_nvm_is_installed
+nvm install --lts
 
-# Java
+# Java etc.
 brew reinstall java11
 brew reinstall maven
 brew reinstall gradle
@@ -50,7 +63,7 @@ brew tap pivotal/tap
 brew reinstall springboot
 
 # Docker etc.
-rm /usr/local/bin/docker
+rm -f /usr/local/bin/docker
 brew reinstall docker
 rm -f /usr/local/bin/kubectl
 brew reinstall kubectl
@@ -75,16 +88,6 @@ append_to_zshrc "source ${basePath}/components/zshrc/aliases/git.sh" 1
 append_to_zshrc "source ${basePath}/components/zshrc/aliases/docker.sh" 1
 
 append_to_zshrc "export GPG_TTY=$\(tty\)"
-
-rm -rf ~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-git clone https://github.com/zsh-users/zsh-autosuggestions ~/.oh-my-zsh/custom}/plugins/zsh-autosuggestions
-update_file_line_in_situ ~/.zshrc 'plugins=(git)' 'plugins=(docker git zsh-autosuggestions)'
-
-ensure_zshrc_correction_is_used
-
-ensure_zshrc_completion_waiting_dots_are_used
-
-ensure_correct_ohmyzsh_theme_is_used "${basePath}/components/ohmyzsh/willgibson.zsh-theme" "willgibson"
 
 fancy_echo "To reload profile now please run...\nsource ~/.zshrc"