+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Install Intellij IDEA
 - Install Postman
 - Install Spotify
+### Changed
+- Updated Homebrew install command
 
 ## [0.1.0] - 2020-06-12
 ### Added