+++ CHANGELOG.md
@@ -5,9 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [0.1.5] - 2020-10-26
 ### Added
 - GitHub Actions pipeline
+- Install VS Code
+- Install Slack
+- Install Microsoft Teams
+- Install Google Chrome
 ### Changed
 - Split up the commands and extract more things to one liners